refactor: token 리팩터링 (infra 계층)
